LayoutTests/fast/events/imagemap-norender-crash.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<body id="body">

<h1 id="header" tabindex=0>start element</h1>
<div><img src="resources/abe.png" alt="test" style="width:679px; height:112px" usemap="#Map"></div>
<div style="display:none"><map name="Map" id="Map"><area shape="rect" coords=coords="5,48,247,97" href="http://www.webkit.org/" target="_blank" id="area1"/></map></div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests tabbing to an image map link where the map might not have a renderer won't crash.");

    // start at the right place
    document.getElementById("header").focus();

    // tab forward
    eventSender.keyDown('\t');
    shouldBe("document.activeElement.id", "'area1'");
</script>

<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
